Add blscfg command support to parse BootLoaderSpec config fragments
The BootLoaderSpec (BLS) defines a scheme where different bootloaders can
share a format for boot items and a configuration directory that accepts
these common configurations as drop-in files.

GRUB2 already has a blscfg modle that can parse the config snippets using
the bls_import command, change it to blscfg and improve the BLS support.