fs: Use 64bit type for filesystem timestamp
Some filesystems nowadays uses 64bit types for timestamps, so, update
grub_dirhook_info struct to use an int64 type to store mtime. This also
updates grub_unixtime2datetime() to receive a 64-bit timestamp
argument and do 64bit-safe divisions.

All the remaining conversion from 32 to 64 should be safe, as 32 to 64
attributions will be implicitly casted. The most crictical part in the
32 to 64bits conversion is on grub_unixtime2datetime() where it needs
to deal with the 64bit type, specially with division in x86_architectures,
so, for that, the grub_divmod64() helper has been used.

These changes enables grub to support dates beyond y2038.

Reviewed-by: Javier Martinez Canillas <javierm@redhat.com>
Signed-off-by: Carlos Maiolino <cmaiolino@redhat.com>